requirements.txt
asgiref==3.8.1
contourpy==1.3.2
cycler==0.12.1
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
et_xmlfile==2.0.0
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
